UPDATE >_ 110
Sarwan executed the development of the Word Page and Paragraph Page screens, incorporating history viewers for each and establishing their respective functionalities, inclusive of an About screen. Furthermore, the implementation included the successful design and construction of the update-check feature and the interfaces for the Word Page and Paragraph Page. This undertaking was completed, and the update was disseminated at 00:14 AM on December 04, 2023.