UPLOAD >_  110
Committed executable file to the repository, marking a pivotal advancement in the project's software deployment. Additionally, initiated a comprehensive upgrade to the project website, enhancing its functionality and overall user experience. This commit reflects a strategic step forward in the ongoing development and improvement of the software system, aligning with our commitment to delivering high-quality and feature-rich solutions to our users.

Co-Authored-By: Sarwan Yadav <92580571+sarwanboi@users.noreply.github.com>